@@ -148,7 +148,7 @@ class _ForgotPasswordView extends State<ForgotPasswordView>
   phoneTextFormFiled() {
     return SizedBox(
       height: Dimens.sixty,
-      width: Get.width,
+      width: MediaQuery.of(context).size.width,
       child: TextFormField(
         controller: phoneController,
         autovalidateMode: AutovalidateMode.onUserInteraction,
