@@ -39,8 +39,8 @@ class _GetStartedView extends State<GetStartedView>
           child: Stack(
             children: [
               SizedBox(
-                width: Get.width,
-                height: Get.height,
+                width: MediaQuery.of(context).size.width,
+                height: MediaQuery.of(context).size.height,
                 child: SvgPicture.asset(
                   AssetConstants.getStartedBg,
                   alignment: Alignment.center,
