@@ -109,41 +109,66 @@ class ProductData {
   String? warranty_type;
   String? refundable;
   String? refend_day;
+  String? billing_mode_of_shipment;
+  String? seo_product_title;
+  String? seo_product_meta;
+  String? targetted_keywords;
+  String? maximum_shipping_cost;
+  String? video_url;
+  String? weight_in_grams;
+  String? length_in_cm;
+  String? breadth_in_cm;
+  String? height_in_cm;
+  String? delivery_charges_for_above_30kg_items;
+  String? status_of_shipment;
+  String? available_pin_code;
 
-  ProductData({
-    this.id,
-    this.vendorId,
-    this.categoryId,
-    this.subCategoryId,
-    this.childCategoryId,
-    this.isPopular,
-    this.isTaxable,
-    this.hsnNo,
-    this.taxValue,
-    this.sameSayDelivery,
-    this.deliveryDay,
-    this.brandId,
-    this.name,
-    this.product_rating,
-    this.description,
-    this.images,
-    this.inStock,
-    this.attributeIds,
-    this.createdAt,
-    this.updatedAt,
-    this.image,
-    this.variations,
-    this.category,
-    this.brand,
-    this.specification,
-    this.cash_on_delivery,
-    this.warranty,
-    this.warranty_month,
-    this.refundable,
-    this.refend_day,
-    this.warranty_type,
-    this.default_image,
-  });
+  ProductData(
+      {this.id,
+      this.vendorId,
+      this.categoryId,
+      this.subCategoryId,
+      this.childCategoryId,
+      this.isPopular,
+      this.isTaxable,
+      this.hsnNo,
+      this.taxValue,
+      this.sameSayDelivery,
+      this.deliveryDay,
+      this.brandId,
+      this.name,
+      this.product_rating,
+      this.description,
+      this.images,
+      this.inStock,
+      this.attributeIds,
+      this.createdAt,
+      this.updatedAt,
+      this.image,
+      this.variations,
+      this.category,
+      this.brand,
+      this.specification,
+      this.cash_on_delivery,
+      this.warranty,
+      this.warranty_month,
+      this.refundable,
+      this.refend_day,
+      this.warranty_type,
+      this.default_image,
+      this.billing_mode_of_shipment,
+      this.seo_product_title,
+      this.seo_product_meta,
+      this.targetted_keywords,
+      this.maximum_shipping_cost,
+      this.video_url,
+      this.weight_in_grams,
+      this.length_in_cm,
+      this.breadth_in_cm,
+      this.height_in_cm,
+      this.delivery_charges_for_above_30kg_items,
+      this.status_of_shipment,
+      this.available_pin_code});
 
   ProductData.fromJson(Map<String, dynamic> json) {
     id = json['id'];
@@ -192,6 +217,20 @@ class ProductData {
     warranty_month = json['warranty_month'].toString();
     refundable = json['refundable'].toString();
     refend_day = json['refend_day'].toString();
+    billing_mode_of_shipment = json['billing_mode_of_shipment'].toString();
+    seo_product_title = json['seo_product_title'].toString();
+    seo_product_meta = json['seo_product_meta'].toString();
+    targetted_keywords = json['targetted_keywords'].toString();
+    maximum_shipping_cost = json['maximum_shipping_cost'].toString();
+    video_url = json['video_url'].toString();
+    weight_in_grams = json['weight_in_grams'].toString();
+    length_in_cm = json['length_in_cm'].toString();
+    breadth_in_cm = json['breadth_in_cm'].toString();
+    height_in_cm = json['height_in_cm'].toString();
+    delivery_charges_for_above_30kg_items =
+        json['delivery_charges_for_above_30kg_items'].toString();
+    status_of_shipment = json['status_of_shipment'].toString();
+    available_pin_code = json['available_pin_code'].toString();
   }
 
   Map<String, dynamic> toJson() {
@@ -232,6 +271,22 @@ class ProductData {
       data['specification'] =
           this.specification!.map((v) => v.toJson()).toList();
     }
+    data['billing_mode_of_shipment'] = billing_mode_of_shipment;
+    data['seo_product_title'] = seo_product_title;
+    data['seo_product_meta'] = seo_product_meta;
+    data['targetted_keywords'] = targetted_keywords;
+    data['maximum_shipping_cost'] = maximum_shipping_cost;
+
+    data['video_url'] = video_url;
+    data['weight_in_grams'] = weight_in_grams;
+    data['length_in_cm'] = length_in_cm;
+    data['breadth_in_cm'] = breadth_in_cm;
+    data['height_in_cm'] = height_in_cm;
+    data['delivery_charges_for_above_30kg_items'] =
+        delivery_charges_for_above_30kg_items;
+    data['status_of_shipment'] = status_of_shipment;
+    data['available_pin_code'] = available_pin_code;
+
     return data;
   }
 }
