@@ -283,7 +283,7 @@ class _LoginWithOTPView extends State<LoginWithOTPView>
   phoneTextFormFiled() {
     return SizedBox(
       height: Dimens.sixty,
-      width: Get.width,
+      width: MediaQuery.of(context).size.width,
       child: TextFormField(
         controller: phoneController,
         autovalidateMode: AutovalidateMode.onUserInteraction,
