@@ -125,7 +125,7 @@ class _DeclineOrderScreenState extends State<DeclineOrderScreen>
                   ),
                   Container(
                     margin: const EdgeInsets.only(left: 16, right: 16, top: 16),
-                    width: Get.width,
+                    width: MediaQuery.of(context).size.width,
                     height: 50,
                     alignment: Alignment.centerLeft,
                     child: Row(
