@@ -1,5 +1,5 @@
 class ApiInterface {
-  void onSuccess(var data) {}
-  void onFailure(var message) {}
+  void onSuccess(dynamic data) {}
+  void onFailure(dynamic message) {}
   void onTokenExpired() {}
 }
